Move testing of containerisation to test_lib.py
